requirements.txt
Flask
Flask-SocketIO
kafka-python